Simulators/CN_WIRED: Tidy up
- Fix "indoors temp =" logging by changing type to int, so that there's no more
  need for + prefix.
- Pressing enter in "Serial port" widget now automatically starts the
  simulation.

+++ Simulators/CN_WIRED/ACSimulator/simulator.cpp
@@ -281,7 +281,7 @@ static void sendStatePacket(QSerialPort* serial)
 }
 
 // Callbacks for the UI
-void setIndoorTemp(uint8_t temp, QSerialPort* serial)
+void setIndoorTemp(int temp, QSerialPort* serial)
 {
     uint8_t response[CNW_PKT_LEN] = {0};
