+++ Simulators/CN_WIRED/ACSimulator/simulator.hpp
@@ -7,7 +7,7 @@ QString openSerial(QSerialPort* serial, const QString& port);
 void closeSerial(QSerialPort* serial);
 void serialRead(QSerialPort* serial);
 
-void setIndoorTemp(uint8_t temp, QSerialPort* serial);
+void setIndoorTemp(int temp, QSerialPort* serial);
 void setPower(bool on, QSerialPort* serial);
 void setPoint(uint8_t temp, QSerialPort* serial);
 void setMode(uint8_t mode, QSerialPort* serial);
